requirements.txt
pyyaml
requests
pandas
yfinance
mysql-connector-python
pymongo
pyspark